don't specify polarization at the level of KetBra trees
The map between experimental angles and dummy angles is established by tracking
the order of interactions independently of specified polarization. Fake integer
polarizations are no longer used to do so. The actual polarizations of the beams
are specified only when geometric_factor or a related quantity has to be
calculated.